Cleanup file listing.
Allow the user to pass a directory on the command line. If not passed, use the provided src/main/resources. Now that we accept input, we can no longer assume that the files can be opened from resources.

Stop passing around an InputStream that will just be converted to a scala.io.Source and use an absolute path instead.

Just for kicks, get rid of the additional list conversion done when getting the file list.